requirements.txt
python-dotenv
discord
asyncio